♻️ Use number index by default
Reason : easier to manipulate than Int

+++ packages/core/src/Char/CharIndexable.ts
@@ -1,7 +1,6 @@
 import { Indexable } from '../Indexable.js';
-import type { Int } from '../Int.js';
 
 export const CharIndexable = Indexable({
   at: (index) => String.fromCodePoint(index),
-  indexOf: (value) => value.codePointAt(0) as Int,
+  indexOf: (value) => value.codePointAt(0),
 });
